Create toggle macro when dropping conditions in hotbar

+++ src/scripts/macros/hotbar.ts
@@ -1,7 +1,7 @@
 import { SkillAbbreviation } from "@actor/creature/data.ts";
 import { StrikeData } from "@actor/data/base.ts";
 import { SKILL_DICTIONARY } from "@actor/values.ts";
-import { EffectPF2e } from "@item";
+import type { ConditionPF2e, EffectPF2e } from "@item";
 import { ChatMessagePF2e } from "@module/chat-message/document.ts";
 import { createSelfEffectMessage } from "@module/chat-message/helpers.ts";
 import { MacroPF2e } from "@module/macro.ts";
@@ -118,7 +118,7 @@ if (a) {
     game.user.assignHotbarMacro(skillMacro ?? null, slot);
 }
 
-export async function createToggleEffectMacro(effect: EffectPF2e, slot: number): Promise<void> {
+export async function createToggleEffectMacro(effect: ConditionPF2e | EffectPF2e, slot: number): Promise<void> {
     const uuid = effect.uuid.startsWith("Actor") ? effect.sourceId : effect.uuid;
     if (!uuid) {
         ui.notifications.error("PF2E.ErrorMessage.CantCreateEffectMacro", { localize: true });
@@ -126,23 +126,32 @@ export async function createToggleEffectMacro(effect: EffectPF2e, slot: number):
     }
 
     const command = `
-const actors = canvas.tokens.controlled.flatMap((token) => token.actor ?? []);
+const actors = Array.from(new Set(canvas.tokens.controlled.flatMap((token) => token.actor ?? [])));
 if (actors.length === 0 && game.user.character) actors.push(game.user.character);
 if (actors.length === 0) {
     return ui.notifications.error("PF2E.ErrorMessage.NoTokenSelected", { localize: true });
 }
 
 const ITEM_UUID = "${uuid}"; // ${effect.name}
-const source = (await fromUuid(ITEM_UUID)).toObject();
-source.flags = mergeObject(source.flags ?? {}, { core: { sourceId: ITEM_UUID } });
-
-for (const actor of actors) {
-    const existing = actor.itemTypes.effect.find((e) => e.flags.core?.sourceId === ITEM_UUID);
-    if (existing) {
-        await existing.delete();
-    } else {
-        await actor.createEmbeddedDocuments("Item", [source]);
+const item = await fromUuid(ITEM_UUID);
+if (item?.type === "condition") {
+    for (const actor of actors) {
+        actor.toggleCondition(item.slug);
     }
+} else if (item?.type === "effect") {
+    const source = item.toObject();
+    source.flags = mergeObject(source.flags ?? {}, { core: { sourceId: ITEM_UUID } });
+
+    for (const actor of actors) {
+        const existing = actor.itemTypes.effect.find((e) => e.flags.core?.sourceId === ITEM_UUID);
+        if (existing) {
+            await existing.delete();
+        } else {
+            await actor.createEmbeddedDocuments("Item", [source]);
+        }
+    }
+} else {
+    ui.notifications.error(game.i18n.format("PF2E.ErrorMessage.ItemNotFoundByUUID", { uuid: ITEM_UUID }));
 }
 `;
     const toggleMacro =
